moving attackResult and AttackOutcome
