+++ pyrit/attacks/__init__.py
@@ -1,7 +1,6 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-
 from pyrit.attacks.base.attack_config import AttackAdversarialConfig, AttackScoringConfig, AttackConverterConfig
 from pyrit.attacks.base.attack_context import (
     AttackContext,
@@ -11,7 +10,6 @@ from pyrit.attacks.base.attack_context import (
     ContextT,
 )
 from pyrit.attacks.base.attack_executor import AttackExecutor
-from pyrit.attacks.base.attack_result import AttackResult, ResultT, AttackOutcome
 from pyrit.attacks.base.attack_strategy import AttackStrategy, AttackStrategyLogAdapter
 from pyrit.attacks.multi_turn.red_teaming import RedTeamingAttack, RTOSystemPromptPaths
 from pyrit.attacks.multi_turn.crescendo import CrescendoAttack
@@ -28,9 +26,6 @@ __all__ = [
     "ConversationSession",
     "ContextT",
     "AttackExecutor",
-    "AttackResult",
-    "ResultT",
-    "AttackOutcome",
     "AttackStrategy",
     "AttackStrategyLogAdapter",
     "RedTeamingAttack",
