+++ pyrit/models/attack_result.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass, field
 from enum import Enum
 from typing import Any, Dict, Optional, TypeVar
 
-from pyrit.models.prompt_request_piece import PromptRequestPiece
+from pyrit.models import PromptRequestPiece 
 from pyrit.models.score import Score
 
 ResultT = TypeVar("ResultT", bound="AttackResult")
