+++ tests/unit/attacks/test_red_teaming.py
@@ -17,7 +17,7 @@ from pyrit.attacks.base.attack_context import (
     ConversationSession,
     MultiTurnAttackContext,
 )
-from pyrit.attacks.base.attack_result import AttackOutcome, AttackResult
+from pyrit.models import AttackOutcome, AttackResult
 from pyrit.attacks.components.conversation_manager import ConversationState
 from pyrit.attacks.multi_turn.red_teaming import RedTeamingAttack, RTOSystemPromptPaths
 from pyrit.exceptions.exception_classes import (
